use tests dirs, benchmarks, use __getattr__
